register, login, logout user logic; create, update, delete user's daily; characters as clent-side request to provide firebase with auth; characters and unitShards from swgoh API are cached to firebase to minimize app's first load, cache logic

+++ pages/index.jsx
@@ -1,62 +1,55 @@
 import { useState, useEffect, useMemo, useCallback } from 'react';
 import {
   addDoc,
-  getDocs,
+  query,
+  where,
   onSnapshot,
   deleteDoc,
   updateDoc,
+  getDoc,
 } from 'firebase/firestore';
-import { Container, Autocomplete, TextField, Box } from '@mui/material';
+import { Container, Box } from '@mui/material';
 
-import { dailiesCollection, getDailiesDocById } from '../src/db';
-import { login, getUnitsList, getMaterialList } from '../src/services';
+import { dailiesCollection } from '../src/db';
+import {
+  getDailiesDocById,
+  getCharactersWithShardsProps,
+} from '../src/services';
 import {
   getOptionsFromCharacters,
-  getUnitShardsData,
   getFirebaseData,
-  normalizeArray,
   getLocationsAndTries,
 } from '../src/helpers';
+
+import { useUser } from '../src/hooks';
 
 import Layout from '../src/components/Layout';
 import DailiesList from '../src/components/DailiesList';
 import AddDaily from '../src/components/AddDaily';
 
-export const getStaticProps = async () => {
-  await login();
-  const charactersArray = await getUnitsList();
-  const materials = await getMaterialList();
-
-  const unitShardsArray = getUnitShardsData(materials);
-
-  const characters = normalizeArray(charactersArray, 'baseId');
-  const unitShards = normalizeArray(unitShardsArray, 'id');
-
-  const dailiesDocs = await getDocs(dailiesCollection);
-  const dailies = dailiesDocs.docs.map(getFirebaseData);
-  return {
-    props: {
-      characters,
-      unitShards,
-      dailies,
-    },
-    revalidate: 3600,
-  };
-};
-
-const Home = props => {
-  const { characters, unitShards } = props;
-  const [dailies, setDailies] = useState(props.dailies);
+const Home = () => {
+  const [dailies, setDailies] = useState([]);
+  const [characters, setCharacters] = useState({});
+  const { uid } = useUser();
 
   useEffect(() => {
-    const dailiesSubscription = onSnapshot(dailiesCollection, data =>
+    (async () => {
+      const characters = await getCharactersWithShardsProps();
+      setCharacters(characters);
+    })();
+  }, []);
+
+  useEffect(() => {
+    const dailiesQuery = query(dailiesCollection, where('uid', '==', uid));
+
+    const dailiesSubscription = onSnapshot(dailiesQuery, data =>
       setDailies(data.docs.map(getFirebaseData)),
     );
 
     return () => {
       dailiesSubscription();
     };
-  }, []);
+  }, [uid]);
 
   const options = useMemo(
     () => getOptionsFromCharacters(characters),
@@ -67,16 +60,23 @@ const Home = props => {
     if (!baseId) return;
     addDoc(dailiesCollection, {
       baseId,
+      uid,
     });
   };
 
-  const deleteDaily = id => {
+  const deleteDaily = async id => {
     const docToDelete = getDailiesDocById(id);
+    const doc = await getDoc(docToDelete);
+    if (doc.data().uid !== uid) return;
+
     deleteDoc(docToDelete);
   };
 
-  const updateDaily = ({ id, idx, payload }) => {
+  const updateDaily = async ({ id, idx, payload }) => {
     const docToUpdate = getDailiesDocById(id);
+    const doc = await getDoc(docToUpdate);
+    if (doc.data().uid !== uid) return;
+
     const prevTries = dailies.find(daily => daily.id === id)?.tries ?? {};
     const tries = {
       ...prevTries,
@@ -87,20 +87,20 @@ const Home = props => {
 
   const getCharacterData = useCallback(
     (baseId, tries = {}) => {
-      const character = characters[baseId];
-      const baseLocations = unitShards[baseId]?.locations;
+      const character = characters[baseId] ?? {};
+      const baseLocations = character?.locations ?? [];
       const locationsAndTries = getLocationsAndTries(baseLocations, tries);
 
       return { ...character, ...locationsAndTries };
     },
-    [characters, unitShards],
+    [characters],
   );
 
   const dailiesList = useMemo(
     () =>
       dailies.map(({ id, baseId, tries }) => ({
-        id,
         ...getCharacterData(baseId, tries),
+        id,
       })),
     [dailies, getCharacterData],
   );
